src/config/chatbot_v2/requirements.txt
# AI Service Packages
anthropic>=0.39.0
google-generativeai>=0.3.0
openai>=1.0.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
